requirements.txt
beautifulsoup4
PyYAML
pytest-playwright
torch
ultralytics
